+++ src/graphics/point.rs
@@ -1,11 +1,11 @@
 #[derive(Copy, Clone)]
 pub struct Point {
-	pub x: i32,
-	pub y: i32
+	pub x: isize,
+	pub y: isize
 }
 
 impl Point {
-	pub fn new(x: i32, y: i32) -> Point {
+	pub fn new(x: isize, y: isize) -> Point {
 		Point { x: x, y: y }
 	}
 }
